Do Not Block Realtime Loop
This is just wrong. The only way to fix it properly is to redesign the software
to actually use ros_control instead of working around it.

If you switch the control type while controllers are running, behavior is undefined now.
Especially for position controllers you will experience broken behavior when switching pwm->torque.

Disabling running controllers, switching the control type and starting the intended new ones takes <300ms on average where the current approach took 8 seconds.

It would be nice to be able to use switch_controllers with start and stop controller lists together and have the control type change under the hood, but as stated above this requires a major redesign I'm not willing to undertake.